mcp-server/pyproject.toml
[project]
name = "dispenser-mcp-server"
version = "0.1.0"
description = "MCP Server for dispenser-ai system capabilities"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
